Change Pitch fields to type int

+++ dragonfly/session/world.go
@@ -369,7 +369,7 @@ func (s *Session) ViewParticle(pos mgl64.Vec3, p world.Particle) {
 	case particle.Note:
 		s.writePacket(&packet.BlockEvent{
 			EventType: int32(pa.Instrument.MagicNumber),
-			EventData: pa.Pitch,
+			EventData: int32(pa.Pitch),
 			Position:  protocol.BlockPos{int32(pos.X()), int32(pos.Y()), int32(pos.Z())},
 		})
 	case particle.HugeExplosion:
@@ -412,7 +412,7 @@ func (s *Session) ViewSound(pos mgl64.Vec3, soundType world.Sound) {
 	switch so := soundType.(type) {
 	case sound.Note:
 		pk.SoundType = packet.SoundEventNote
-		pk.ExtraData = (int32(so.Instrument.MagicNumber) << 8) | so.Pitch
+		pk.ExtraData = (int32(so.Instrument.MagicNumber) << 8) | int32(so.Pitch)
 	case sound.DoorCrash:
 		s.writePacket(&packet.LevelEvent{
 			EventType: packet.EventSoundDoorCrash,
